Add files via upload
Facial recognition is a way of identifying or confirming an individual’s identity using their 
face. Facial recognition systems can be used to identify people in photos, videos, or in realtime.
Facial recognition is a category of biometric security. Other forms of biometric software 
include voice recognition, fingerprint recognition, and eye retina or iris recognition. The 
technology is mostly used for security and law enforcement, though there is increasing interest 
in other areas of use.
The principal use of face recognition technology continues to be security solutions. Facial 
recognition is recognized as one of the most accurate and easy ways for establishing individual 
identity across a wide range of sectors.
We have prepared our project with the help of Raspberry pi Using python, we make a face 
recognition system.
While humans can recognize faces without much effort, facial recognition is a challenging 
pattern recognition problem in computing. Facial recognition systems attempt to identify a 
human face, which is three-dimensional and changes in appearance with lighting and facial 
expression, based on its two-dimensional image. To accomplish this computational task, facial 
recognition systems perform four steps. First face detection is used to segment the face from 
the image background. In the second step the segmented face image is aligned to account for 
face pose, image size and photographic properties, such as illumination and grayscale. The 
purpose of the alignment process is to enable the accurate localization of facial features in the 
third step, the facial feature extraction. Features such as eyes, nose and mouth are pinpointed 
and measured in the image to represent the face. The so established feature vector of the face 
is then, in the fourth step, matched against a database of faces